move methods to their respective places
