+++ client/client.go
@@ -5,7 +5,6 @@ import (
 	"crypto/tls"
 	"fmt"
 	"net/http"
-	"strings"
 
 	"github.com/gomicro/trust"
 	"github.com/google/go-github/v56/github"
@@ -55,40 +54,4 @@ func New(tkn string) (*Client, error) {
 		ghClient: github.NewClient(oauth2.NewClient(ctx, ts)),
 		rate:     rl,
 	}, nil
-}
-
-func (c *Client) GetLogins(ctx context.Context) ([]string, error) {
-	logins := []string{}
-
-	user, _, err := c.ghClient.Users.Get(ctx, "")
-	if err != nil {
-		if _, ok := err.(*github.RateLimitError); ok {
-			return nil, fmt.Errorf("github: hit rate limit")
-		}
-
-		return nil, fmt.Errorf("get user: %w", err)
-	}
-
-	logins = append(logins, strings.ToLower(user.GetLogin()))
-
-	opts := &github.ListOptions{
-		Page:    0,
-		PerPage: 100,
-	}
-
-	orgs, _, err := c.ghClient.Organizations.List(ctx, "", opts)
-	if err != nil {
-		if _, ok := err.(*github.RateLimitError); ok {
-			return nil, fmt.Errorf("github: hit rate limit")
-		}
-
-		return nil, fmt.Errorf("list orgs: %w", err)
-	}
-
-	for i := range orgs {
-		o := orgs[i].GetLogin()
-		logins = append(logins, strings.ToLower(o))
-	}
-
-	return logins, nil
 }
